refactor(image-utils): remove `main` export
We don't actually want this exported from the library.
